DT-291 Added services page testing functionality
