+++ frontend/fresh-cuts/src/scenes/services/index.jsx
@@ -36,7 +36,7 @@ const Services = () => {
               <Grid item xs={12} sm={6} md={4} key={service.id}>
                 <Card style={{ height: "100%", backgroundColor: "#f2f2f2", display: 'flex', flexDirection: 'column' }}>
                   <CardContent>
-                    <Typography variant="h6" component="div">
+                    <Typography variant="h6" component="div" aria-label={service.name}>
                       {service.name}
                     </Typography>
                     {service.description && (
@@ -54,13 +54,16 @@ const Services = () => {
                   <div style={{ marginTop: 'auto', textAlign: 'right', padding: '15px' }}>
                     <Button
                       variant="contained"
+                      aria-label={`REQUEST${service.id}`}
+                      href={`http://localhost:3000/booking`}
                       style={{
                         backgroundColor: "#E95252",
                         color: "white",
                         transition: "background-color 0.3s",
                       }}
                     >
-                      <Link style={{textDecoration: "none", color: "white"}} to="/booking">Request</Link>
+                      {/*<Link style={{textDecoration: "none", color: "white"}} to="/booking">Request</Link>*/}
+                      Request
                     </Button>
                   </div>
                 </Card>
