+++ frontend/fresh-cuts/src/scenes/staff/staffInsta.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import {render, screen, waitFor} from '@testing-library/react';
+import Staff from './index';
+import userEvent from '@testing-library/user-event';
+
+for(let i = 0; i < 7; i++) {
+  test(`handle instagram click ${i}`, async () => {
+    render(<Staff/>);
+    await waitFor(() => userEvent.click(screen.getByLabelText(`INSTA${i}`)));
+  })
+}
